GIX-1823: Converter agg data to SnsSummary (#3138)
# Motivation

Final motivation: Stop using get_state. But on the way, clean the
unnecessary complexity of snsQueryStore.

In this PR: Introduce the converter from sns aggreagator data to
SnsSummary.

The SnsSummary type is used in the `snsSummariesStore` to return a list
of SnsSummary.

After this PR, we can enable using the aggregator data instead of
snsQueryStore.

Then, we can remove snsQueryStore.

# Changes

* New sns aggregator converter `convertDtoToSnsSummary`. Some of the
utils are similar to utils in sns.utils.ts. I didn't use those on
purpose because I want to remove them along the snsQueryStore.

# Tests

* Add a test for the new converter `convertDtoToSnsSummary`.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet worth an entry.