GIX-1823: New feature flag ENABLE_SNS_AGGREGATOR_STORE (#3142)
# Motivation

New feature flag to test using the new snsAggregatorStore instead of
snsQueryStore in mainnet: "ENABLE_SNS_AGGREGATOR_STORE".

This is a sensitive store that collects all the SNS project data and
serves it to the components. Therefore, we want to be sure that
everything is working before going to prod.

Specially since the aggregator data is not typed and there could be
unexpected bugs.

Note: The feature flag is NOT yet USED. This comes in another PR.

# Changes

* Add new feature flag ENABLE_SNS_AGGREGATOR_STORE as "false"
everywhere.

# Tests

* Add new feature flag in env vars tests.

# Todos

- [x] Add entry to changelog (if necessary).