Fx: Remove unused converters (#3155)
# Motivation

We had a bug in `main` before production because a sns aggregator
converter that wasn't used was creating problems.

The issue with the converter has been fixed. Yet, we don't need to do
some conversions to store the data in snsQueryStore.

In this PR, I removed the fields that the aggregator exposes, but we
don't use for the snsQueryStore.

# Changes

* Remove fields `swap_params`, `init` and `lifecycle` from CachedSns
type and converter. These fields are not used to store data in
snsQueryStore.

# Tests

* Remove a test case that was checking a removed field.

# Todos

- [x] Add entry to changelog (if necessary).