Chore: Add Launchpad test with prod data (#3160)
# Motivation

We had a bug because the type of the aggregator in ths NNS Dapp was not
correct and when it tried to convert a certain field it failed.

In this PR, I try to add a test that renders the Launchpad with prod
data from the aggregator. I then check that there are committed projects
and no errors.

# Changes

* New test  in workflows/Launchpad/ProdLaunchpad.spec.ts.
* New test component in workflows/Launchpad/LaunchpadWithLayout.svelte.
This is used to trigger the same flows as the svelte app. In this
scenario, the call to the SNS aggregator.

# Tests

Only test changes.

# Todos

- [ ] Add entry to changelog (if necessary).