Rustdoc - fix (#3170)
# Motivation
Rust code documentation can be parsed and displayed in a nice, readable
way by rustdoc.

The sns aggregator documentation is available in this way, but
unfortunately most of the nns-dapp code is inaccessible, making it
unnecessarily to read.

# Changes
* Make the nns-dapp code accessible to the nns-dapp lib, which is the
entry point for rustdoc.
* Note: `rustdoc` _does_ have limited support for documentation of
binaries, however it is not good.
* Fix rustdoc links
* Test the rustdoc in CI.
* Add two utility scripts to make it easy to generate documentation
without retyping a bunch of flags.
* Note: `rustdoc` flags such as `--open` are passed through the script.
* Performed a quick pass checking spelling and formatting. E.g. code
`should be in backticks`.

# Tests
*  `cargo doc` is now tested in CI.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>