Fix: More SNS aggregator optional types (#3168)
# Motivation

We realized that some of our SNS aggregator types were not correct. It
seems that the changes were not exhaustive and there are more fields
that can still be optional.

I based my findings on the actual types of the canisters. Where the
response said `[] | [T]`, I changed it to `null | T`.

# Changes

* Change value of some fields in SNS aggregator to make them optional.
* Fix converters after change.

# Tests

* Test the new functionality when converting the summary that might
return undefined if swap params are not defined.
* Add a test for `convertNervousFuncttion` (typo will be addressed in
another PR)

# Todos

- [ ] Add entry to changelog (if necessary).
Already covered by a changelog entry.