Avoid visual loading glitch (#3186)
# Motivation

This login glitch has probably always been there since the route are
public but, because the "not logged in" pages were bare minimal, it did
not surfaces visually. Adding more content in these page makes it more
obvious.

# Changes

- wait for the authentication to be loaded before rendering sign-in or
not content

# Notes

This is my go to approach in any dapps on the IC, don't know why I never
implemented it in NNS Dapp 🤷‍♂️.

# Screenshots

Glitch:


![glitch](https://github.com/dfinity/nns-dapp/assets/16886711/4ba639fe-5ee0-43d2-9a95-ae7a1b33043f)

No glitch:


![noglitch](https://github.com/dfinity/nns-dapp/assets/16886711/a4a7a56d-b7a5-4c3e-a85e-2b4472eaeb83)