Chore: Upgrade gix components (#3211)
# Motivation

We added an extra `div` wrapper to PageBanner because there was an issue
with grid container inside a flex.

This wrapper was added in the `PageBanner` inside gix-components to
avoid putting it in the parent.

In this PR, we upgrade gix-components.

Next PR, remove the wrapper in NNS dapp.

# Changes

* Upgrade gix-components

# Tests

Not necessary

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.