Fix: Header of accounts page is empty after login (#3212)
# Motivation

The header of the page after logging in was empty. Like it is for
non-logged in users. See the screenshot attached.

Yet, the header should be present for logged in users. As I show in the
screenshot after refreshing the page being logged in.

The problem was that the SignInAccounts was setting the layoutStore
header to empty string. When the user came back from II, the LayoutList
didn't get another rerender because it was the same title.

Solution, do not set the header in two places. Move the responsibility
of not showing the header when not logged in to the LayoutList.

This small issue appeared also for Neurons and Canisters.

Note that the title set in SignInAccounts was not the same as in
LayoutList for accounts page. I used the same for consistency. This is
why I neede to change the expected title in the e2e tests.

# Changes

* Remove setting the title in SignInXxx pages.
* Add a prop to `LayoutList`: `showHeader`, which is used by the layouts
to set the header to empty string if the user is not logged in.
* Use the new prop in the accounts, neuron and canister layouts.

# Tests

* Add a test file for LayoutList.
* Add a test file for accounts layout.
* Add a test file for neurons layout.
* Add a test file for canisters layout.
* Fix title in e2e tests.
* PENDING: Fix screenshot testing

# Todos

- [x] Add entry to changelog (if necessary).