Prevent deep linking ckTESTBTC with only ckBTC enabled (#3207)
# Motivation

I was looking at the test because its success depended on the order in
which tests were run.
But then I realized the test could actually fail while it shouldn't
because `ENABLE_CKTESTBTC` is `true`.
Then I realized that the code that checks if the ckBTC or ckTESTBTC
universe should be enabled doesn't distinguish between the two feature
flags.

# Changes

1. Instead of allowing both universes if either flag is enabled, only
allow each universe if its flag is enabled.
2. In `beforeEach` reset `overrideFeatureFlagsStore` to make sure each
test has a clean slate.
3. Test that ckBTC is not available with `ENABLE_CKTESTBTC` enabled.
4. Corresponding tests for ckTESTBTC.

# Tests

Manually tested that the ckBTC universe can still be selected when the
feature flag is enabled.

# Todos

- [ ] Add entry to changelog (if necessary).
not worth it