Rename MockAccountsDb to AccountsDbAsMap (#3219)
# Motivation
The first production implementation of the accounts DB trait will be the
status quo: Accounts are stored in a hash map.

We actually already have an implementation of the accounts db trait as a
map in the mock, so we can just rename it and move it out of the tests
file.

# Changes
* Rename `MockAccountsDb` to `AccountsDbAsMap`.

# Tests
* The existing tests have been updated.

# Todos

- [ ] Add entry to changelog (if necessary).
- Not needed, as this is minor and doesn't make any user-facing changes.

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>