E2E: Only toggle the address select if necessary (#3239)
# Motivation

There is a TODO in
`frontend/src/tests/page-objects/SelectDestinationAddress.page-object.ts`
to only toggle the input method if necessary.
I'm working on a PR to reuse the `SelectDestinationAddress` component in
the `DisburseNnsNeuronModal` component and for that's it's useful to
finally do this TODO.

# Changes

1. Fix the type of `isChecked`. There is no reason for it to return
`null`.
2. Add functionality to
`frontend/src/tests/page-objects/Toggle.page-object.ts` to check the
value and set a specific value.
3. Add functions to
`frontend/src/tests/page-objects/SelectDestinationAddress.page-object.ts`
to specifically enable select input or text input.
4. Specifically choose select input before selecting an account.

# Tests

test changes only

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary