Add an iterator over values (#3247)
# Motivation
The accounts store tests include an iterator over values. We will need
that in account databases, if we are to keep that test in the current
form. There seems no harm in having such an iterator so just add it.

# Changes
* Add an iterator over accounts.

# Tests
- This PR includes a simple test that inserts a few values and verifies
that the account IDs from the iterator are as expected.
- The iterator will be used in this test:
https://github.com/dfinity/nns-dapp/blob/main/rs/backend/src/accounts_store/tests.rs#L1378

# Todos

- [ ] Add entry to changelog (if necessary).
I think this is not necessary; this is a small PR currently affecting
tests only and has no user-facing changes.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>