GIX-1823: Use aggregator ns functions (#3242)
# Motivation

The service `loadSnsNervousSystemFunctions` checked the dynamic derived
store `createSnsNsFunctionsProjectStore` to check whether it should load
more nervous functions or not.

If that store has data, then call to get the functions and load then in
snsFunctionsStore is skipped.

That means, that we can't rely on the data in snsFunctionsStore, because
it might not be populated.

At the moment, this is working because after getting the aggregator
data, we also load the data in snsFunctionsStore. But, just like we stop
loading snsQueryStore, we will also stop loading snsFunctionsStore.

In this PR, I changed the usage of snsFunctionsStore for the dynamic
derived store `createSnsNsFunctionsProjectStore`.

# Changes

* `` accepts rootCanisterId undefined or null. This make it easy to be
used in components.
* Use `createSnsNsFunctionsProjectStore` in
SnsProposalSystemInfoSection.
* Use `createSnsNsFunctionsProjectStore` in FollowSnsNeuronModal.
* Use `createSnsNsFunctionsProjectStore` in SnsProposals.
* Use `createSnsNsFunctionsProjectStore` in sns vote registration
services.

# Tests

* Tests don't need to change `createSnsNsFunctionsProjectStore` uses
also snsFunctionsStore and tests already populated that store.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.