GIX-1760: Fix spacings nns neuron detail (#3264)
# Motivation

The different sections of NnsNeuronDetails page had the vertical spacing
set in different places. This made it easy to have a different spacing
between them.

In this PR, all the spacing is set in the parent and the children have
no vertical spacing.

# Changes

* Add `noMargin` prop to CardInfo.
* Add `noMargin` prop to SkeletonCard.
* All sections and separators in NnsNeuronDetails are siblings in the
html. Just like it looks like in the UI.
* Use the `noMargin` prop and remove the separator in
NeuronFollowingCard, NeuronProposalsCard, NeuronVotingHistoryCard and
NnsNeuronHotkeysCard.
* Moved the testnet only MakeProposalsCard above the
NeuronVotingHistoryCard to have the beginning like the mainnet. I didn't
move it below NeuronVotingHistoryCard because this one has an infinite
scroll.

# Tests

Only CSS changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary. No UI nor UX change.