GIX-1841: SNS Project "finalizing" status (#3263)
# Motivation

Sns projects perform a finalization process after the swap finishes.
This process might take a few hours. During this time, the users don't
get the neurons yet, even though the project has finished.

To improve the UX, we want to show a different status in the project
detail page during this process takes place.

In this PR, the finalization information is loaded when a user visits
the project detail page and also after participation.

# Changes

* Use new store `isFinalizingStore` in ProjectStatus.svelte to show the
new status or the lifecycle status.
* New service `loadSnsFinalizationStatus` to fetch and load the
finalization in a store.
* Use the new service `loadSnsFinalizationStatus` in ProjectDetai page
when the project is committed.

# Tests

* Test new functionality in ProjectStatus
* Test that finalization status is shown after participation in
ProjectDetail.spec
* Test new service `loadSnsFinalizationStatus`.

# Todos

- [x] Add entry to changelog (if necessary).