GIX-1761: Align spacings sns neuron detail (#3267)
# Motivation

The different sections of SnsNeuronDetails page had the vertical spacing
set in different places. This made it easy to have a different spacing
between them.

In this PR, all the spacing is set in the parent and the children have
no vertical spacing.

# Changes

* All sections and separators in SnsNeuronDetails are siblings in the
html. Just like it looks like in the UI.
* Use the noMargin prop and remove the separator in
SnsNeuronPermissionsCard, SnsNeuronFollowingCard, SnsNeuronHotkeysCard.

# Tests

Only CSS changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.